feat: PinBoard -> PinItem 컴포넌트로 분리
